+++ backend/src/main/java/com/blsa/ezilog/dao/CategoryDao.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.blsa.ezilog.model.category.Category;
 
-public interface CategoryDao extends JpaRepository<Category, BigInteger> {
+public interface CategoryDao extends JpaRepository<Category, Long> {
     
     
     Optional<Category> findCategoryByName(String name);
